Group all Quill modifications under custom theme :tada:
