+++ client/components/content-elements/tce-html/edit/modules/image-embed.js
@@ -1,71 +1,13 @@
-import './image-embed.scss';
-import { Quill } from 'vue-quill-editor';
-import Tooltip from '../ui/tooltip';
+import createImageEmbedTooltip from '../ui/image-embed-tooltip';
 
-const Keyboard = Quill.import('modules/keyboard');
+export default Quill => class ImageEmbed {
+  static NAME = 'imageEmbed';
 
-class EmbedTooltip extends Tooltip {
-  static TEMPLATE = `
-    <div class="controls">
-      <label>Enter image url:</label>
-      <input class="url" type="text">
-      <a class="action">Save</a>
-    </div>`;
-
-  constructor(quill, bounds, { spacing = 0 } = {}) {
-    super(quill, bounds);
-    this.padding = ' '.repeat(spacing);
-    this.textbox = this.root.querySelector('input.url');
-    this.btnAction = this.root.querySelector('a.action');
-    this.listen();
-  }
-
-  listen() {
-    this.btnAction.addEventListener('click', () => this.embedImage());
-    this.textbox.addEventListener('keydown', e => {
-      if (Keyboard.match(e, 'enter')) {
-        this.embedImage();
-        e.preventDefault();
-      } else if (Keyboard.match(e, 'escape')) {
-        this.hide();
-        e.preventDefault();
-      }
-    });
-  }
-
-  show() {
-    super.show();
-    this.root.classList.add('ql-image-embed');
-    this.textbox.select();
-  }
-
-  embedImage() {
-    const url = this.textbox.value;
-    if (!url) return;
-    const { quill } = this;
-    const range = quill.getSelection(true);
-    if (range) {
-      const { padding } = this;
-      const source = Quill.sources.USER;
-      let offset = range.index + range.length;
-      quill.insertText(offset, padding, source);
-      offset += padding.length;
-      quill.insertEmbed(offset, 'image', url, source);
-      offset += 1;
-      quill.insertText(offset, padding, source);
-      offset += padding.length;
-      quill.setSelection(offset, source);
-    }
-    this.textbox.value = '';
-    this.hide();
-  }
-}
-
-export default class ImageEmbed {
   constructor(quill, options = {}) {
     this.quill = quill;
     quill.tooltips = quill.tooltips = {};
     const bounds = quill.options.bounds;
-    quill.tooltips.imageEmbed = new EmbedTooltip(quill, bounds, options);
+    const ImageEmbedTooltip = createImageEmbedTooltip(Quill);
+    quill.tooltips.imageEmbed = new ImageEmbedTooltip(quill, bounds, options);
   }
-}
+};
