+++ client/components/content-elements/tce-html/edit/theme.js
@@ -1,32 +1,37 @@
-import ColorPicker from './ui/color-picker';
-import { Quill } from 'vue-quill-editor';
+import createColorPicker from './ui/color-picker';
+import createImageEmbed from './modules/image-embed';
 
-const SnowTheme = Quill.import('themes/snow');
+export default Quill => {
+  const ImageEmbed = createImageEmbed(Quill);
+  Quill.register(`modules/${ImageEmbed.NAME}`, ImageEmbed, true);
 
-export default class CustomTheme extends SnowTheme {
-  static DEFAULTS = Object.assign({}, SnowTheme.DEFAULTS);
+  return class CustomTheme extends Quill.import('themes/snow') {
+    static DEFAULTS = Object.assign({}, super.DEFAULTS);
+    static NAME = 'tailor';
 
-  buildPickers(selects, icons) {
-    selects = selects.slice();
-    const background = remove(selects, it => it.classList.contains('ql-background'));
-    const color = remove(selects, it => it.classList.contains('ql-color'));
-    super.buildPickers(selects, icons);
-    if (background) {
-      this.pickers.push(new ColorPicker(this.quill, {
-        type: 'background',
-        select: background,
-        label: icons.background
-      }));
+    buildPickers(selects, icons) {
+      selects = selects.slice();
+      const background = remove(selects, it => it.classList.contains('ql-background'));
+      const color = remove(selects, it => it.classList.contains('ql-color'));
+      super.buildPickers(selects, icons);
+      const ColorPicker = createColorPicker(Quill);
+      if (background) {
+        this.pickers.push(new ColorPicker(this.quill, {
+          type: 'background',
+          select: background,
+          label: icons.background
+        }));
+      }
+      if (color) {
+        this.pickers.push(new ColorPicker(this.quill, {
+          type: 'color',
+          select: color,
+          label: icons.color
+        }));
+      }
     }
-    if (color) {
-      this.pickers.push(new ColorPicker(this.quill, {
-        type: 'color',
-        select: color,
-        label: icons.color
-      }));
-    }
-  }
-}
+  };
+};
 
 function remove(arr, cb) {
   const index = arr.findIndex(cb);
